requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faster-whisper @ git+https://github.com/SYSTRAN/faster-whisper
python-multipart==0.0.6
websockets==12.0
numpy==1.24.3
python-dotenv==1.0.0